# openapi.yaml
openapi: 3.2.0
info:
  title: Train Travel API
  description: |
    API for finding and booking train trips across Europe, documented with OpenAPI 3.2.0 specification.
  version: 1.3.0
  contact:
    name: Train Support
    url: https://example.com/support
    email: support@example.com
  license:
    name: Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International
    identifier: CC-BY-NC-SA-4.0
  x-feedbackLink:
    label: Submit Feedback
    url: https://github.com/bump-sh-examples/train-travel-api/issues/new
servers:
  - url: https://train-travel-mock.train-travel-mock.workers.dev
    name: Mock Server
    description: |
      A stateful mock server for testing and development purposes. Book a trip end to end:
      ids returned by one call work in the next. Any `Authorization` header value gives you
      a private data space (none shares a public one); data expires after 24 hours idle.
    x-internal: false

  - url: https://api.example.com
    name: Production
    description: The production server for the Train Travel API
    x-internal: false
...
security:
  - OAuth2:
      - read
...
paths:
...
  /bookings:
    get:
      operationId: get-bookings
      summary: List existing bookings
      description: Returns a list of all trip bookings by the authenticated user.
      tags:
        - Bookings
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: A list of bookings
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            RateLimit:
              $ref: '#/components/headers/RateLimit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Wrapper-Collection'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Booking'
                  - properties:
                      links:
                        allOf:
                          - $ref: '#/components/schemas/Links-Self'
                          - $ref: '#/components/schemas/Links-Pagination'
              examples:
                GetBookingsResponse200:
                  $ref: '#/components/examples/GetBookingsResponse200ApplicationJson'

            application/xml:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Wrapper-Collection'
                  - properties:
                      data:
                        type: array
                        xml:
                          name: bookings
                          wrapped: true
                        items:
                          $ref: '#/components/schemas/Booking'
                  - properties:
                      links:
                        allOf:
                          - $ref: '#/components/schemas/Links-Self'
                          - $ref: '#/components/schemas/Links-Pagination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      operationId: create-booking
      summary: Create a booking
      description: >
        A booking is a temporary hold on a trip. It is not confirmed until the
        payment is processed: its `status` is `pending_payment` until then, and the
        hold ends at `expires_at`. Booking a train that has already departed, or
        bringing a bicycle or a dog on a trip that does not allow them, is a
        conflict.
      tags:
        - Bookings
      security:
        - OAuth2:
            - write
      requestBody:
        description: Booking details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
            examples:
              RequestBodyBooking:
                $ref: '#/components/examples/RequestBodyBookingApplicationJson'
          application/xml:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Booking successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Booking'
                  - properties:
                      links:
                        $ref: '#/components/schemas/Links-Self'
              examples:
                CreateBookingResponse201:
                  $ref: '#/components/examples/CreateBookingResponse201ApplicationJson'

            application/xml:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Booking'
                  - properties:
                      links:
                        $ref: '#/components/schemas/Links-Self'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bookings/{bookingId}:
    parameters:
      - name: bookingId
        in: path
        required: true
        description: The ID of the booking to retrieve.
        schema:
          type: string
          format: uuid
        example: 1725ff48-ab45-4bb5-9d02-88745177dedb
    get:
      summary: Get a booking
      description: Returns the details of a specific booking.
      operationId: get-booking
      tags:
        - Bookings
      responses:
        '200':
          description: The booking details
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            RateLimit:
              $ref: '#/components/headers/RateLimit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Booking'
                  - properties:
                      links:
                        $ref: '#/components/schemas/Links-Self'
              examples:
                GetBookingResponse200:
                  $ref: '#/components/examples/GetBookingResponse200ApplicationJson'
            application/xml:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Booking'
                  - properties:
                      links:
                        $ref: '#/components/schemas/Links-Self'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a booking
      description: Deletes a booking, cancelling the hold on the trip.
      operationId: delete-booking
      security:
        - OAuth2:
            - write
      tags:
        - Bookings
      responses:
        '204':
          description: Booking deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bookings/{bookingId}/payment:
    parameters:
      - name: bookingId
        in: path
        required: true
        description: The ID of the booking to pay for.
        schema:
          type: string
          format: uuid
        example: 1725ff48-ab45-4bb5-9d02-88745177dedb
    post:
      summary: Pay for a Booking
      description: >
        A payment is an attempt to pay for the booking, which will confirm the
        booking for the user and enable them to get their tickets: once it
        succeeds, the booking's `status` becomes `confirmed`. A failed payment
        leaves the booking pending, so it can be retried until the hold expires.
        Paying a booking that is already confirmed returns the existing payment.
      operationId: create-booking-payment
      tags:
        - Payments
      requestBody:
        description: Payment details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingPayment'
            examples:
              Card:
                summary: Card Payment
                dataValue:
                  amount: 49.99
                  currency: gbp
                  source:
                    object: card
                    name: J. Doe
                    number: '4242424242424242'
                    cvc: '123'
                    exp_month: 12
                    exp_year: 2025
                    address_line1: 123 Fake Street
                    address_line2: 4th Floor
                    address_city: London
                    address_country: gb
                    address_post_code: N12 9XX
              Bank:
                summary: Bank Account Payment
                dataValue:
                  amount: 100.5
                  currency: gbp
                  source:
                    object: bank_account
                    name: J. Doe
                    number: '00012345'
                    sort_code: '000123'
                    account_type: individual
                    bank_name: Starling Bank
                    country: gb
      responses:
        '200':
          description: Payment successful
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            RateLimit:
              $ref: '#/components/headers/RateLimit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BookingPayment'
                  - properties:
                      links:
                        $ref: '#/components/schemas/Links-Booking'
              examples:
                Card:
                  summary: Card Payment
                  dataValue:
                    id: 2e3b4f5a-6b7c-8d9e-0f1a-2b3c4d5e6f7a
                    amount: 49.99
                    currency: gbp
                    source:
                      object: card
                      name: J. Doe
                      number: '************4242'
                      exp_month: 12
                      exp_year: 2025
                      address_country: gb
                      address_post_code: N12 9XX
                    status: succeeded
                    links:
                      booking: https://api.example.com/bookings/1725ff48-ab45-4bb5-9d02-88745177dedb
                Bank:
                  summary: Bank Account Payment
                  dataValue:
                    id: 2e3b4f5a-6b7c-8d9e-0f1a-2b3c4d5e6f7a
                    amount: 100.5
                    currency: gbp
                    source:
                      object: bank_account
                      name: J. Doe
                      account_type: individual
                      number: '*********2345'
                      sort_code: '000123'
                      bank_name: Starling Bank
                      country: gb
                    status: succeeded
                    links:
                      booking: https://api.example.com/bookings/1725ff48-ab45-4bb5-9d02-88745177dedb
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    page:
      name: page
      in: query
      description: The page number to return
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    limit:
      name: limit
      in: query
      description: The number of items to return per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
  securitySchemes:
    OAuth2:
      type: oauth2
      description: |
        OAuth 2.0 following RFC8725 best practices, with Authorization Code
        and Device Authorization grants.

        On the mock server, no token is required: any value you enter gives you
        a private sandbox, and without one you share a public sandbox.
      oauth2MetadataUrl: https://example.com/.well-known/oauth-authorization-server
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/oauth/token
          scopes:
            read: Read access
            write: Write access
        deviceAuthorization:
          deviceAuthorizationUrl: https://example.com/oauth/device
          tokenUrl: https://example.com/oauth/token
          scopes:
            read: Read access
            write: Write access
  schemas:
...
    Links-Self:
      description: The link to the current resource.
      type: object
      properties:
        self:
          type: string
          format: uri
...
    Links-Pagination:
      description: Links to the next and previous pages of a paginated response.
      type: object
      properties:
        next:
          type: string
          format: uri
        prev:
          type: string
          format: uri
    Problem:
      description: A problem detail object as defined in RFC 7807.
      type: object
      xml:
        name: problem
        namespace: urn:ietf:rfc:7807
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type
          examples:
          - https://example.com/probs/out-of-credit
        title:
          type: string
          description: A short, human-readable summary of the problem type
          examples:
          - You do not have enough credit.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          examples:
          - Your current balance is 30, but that costs 50.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem
          examples:
          - /account/12345/msgs/abc
        status:
          type: integer
          description: The HTTP status code
          examples:
          - 400
...
    Booking:
      description: A booking for a train trip.
      type: object
      required:
        - trip_id
        - passenger_name
      xml:
        name: booking
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the booking
          readOnly: true
          examples:
            - 1725ff48-ab45-4bb5-9d02-88745177dedb
        trip_id:
          type: string
          format: uuid
          description: Identifier of the booked trip
          examples:
            - ea399ba1-6d95-433f-92d1-83f67b775594
        passenger_name:
          type: string
          description: Name of the passenger
          examples:
            - John Doe
        has_bicycle:
          type: boolean
          description: Indicates whether the passenger has a bicycle.
        has_dog:
          type: boolean
          description: Indicates whether the passenger has a dog.
        status:
          type: string
          description: >
            Where the booking stands: `pending_payment` while the trip is on hold,
            `confirmed` once paid, or `expired` when the hold ended without a
            successful payment.
          enum:
            - pending_payment
            - confirmed
            - expired
          readOnly: true
        price:
          type: number
          description: The fare to pay for this booking.
          readOnly: true
          examples:
            - 50
        currency:
          type: string
          description: Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html) of the fare, in lowercase.
          enum:
            - bam
            - bgn
            - chf
            - eur
            - gbp
            - nok
            - sek
            - try
          readOnly: true
        expires_at:
          type: string
          format: date-time
          description: When the hold on the trip ends if the booking is not paid.
          readOnly: true
    Wrapper-Collection:
      description: This is a generic request/response wrapper which contains both data and links which serve as hypermedia controls (HATEOAS).
      type: object
      properties:
        data:
          description: The wrapper for a collection is an array of objects.
          type: array
          items:
            type: object
        links:
          description: A set of hypermedia links which serve as controls for the client.
          type: object
          readOnly: true
      xml:
        name: data
...
    Links-Booking:
      description: The link to the booking resource.
      type: object
      properties:
        booking:
          type: string
          format: uri
          examples:
            - https://api.example.com/bookings/1725ff48-ab45-4bb5-9d02-88745177dedb
...
  headers:
    Cache-Control:
      description: |
        The Cache-Control header communicates directives for caching mechanisms in both requests and responses.
        It is used to specify the caching directives in responses to prevent caches from storing sensitive information.
      schema:
        type: string
        description: A comma-separated list of directives as defined in [RFC 9111](https://www.rfc-editor.org/rfc/rfc9111.html).
        examples:
          - max-age=3600
          - max-age=604800, public
          - no-store
          - no-cache
          - private
    RateLimit:
      description: |
        The RateLimit header communicates quota policies. It contains a `limit` to
        convey the expiring limit, `remaining` to convey the remaining quota units,
        and `reset` to convey the time window reset time.
      schema:
        type: string
        examples:
          - limit=10, remaining=0, reset=10
    Retry-After:
      description: |
        The Retry-After header indicates how long the user agent should wait before making a follow-up request.
        The value is in seconds and can be an integer or a date in the future.
        If the value is an integer, it indicates the number of seconds to wait.
        If the value is a date, it indicates the time at which the user agent should make a follow-up request.
      schema:
        type: string
      examples:
        integer:
          value: '120'
          summary: Retry after 120 seconds
        date:
          value: 'Fri, 31 Dec 2021 23:59:59 GMT'
          summary: Retry after the specified date
  responses:
    BadRequest:
      description: Bad Request
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            BadRequest:
              summary: A bad request response object
              dataValue:
                type: https://example.com/errors/bad-request
                title: Bad Request
                status: 400
                detail: The request is invalid or missing required parameters.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/bad-request
            title: Bad Request
            status: 400
            detail: The request is invalid or missing required parameters.
    Conflict:
      description: Conflict
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            Conflict:
              summary: A conflict response object
              dataValue:
                type: https://example.com/errors/conflict
                title: Conflict
                status: 409
                detail: There is a conflict with an existing resource.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/conflict
            title: Conflict
            status: 409
            detail: There is a conflict with an existing resource.
    Forbidden:
      description: Forbidden
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            Forbidden:
              summary: A forbidden response object
              dataValue:
                type: https://example.com/errors/forbidden
                title: Forbidden
                status: 403
                detail: Access is forbidden with the provided credentials.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: Access is forbidden with the provided credentials.
    InternalServerError:
      description: Internal Server Error
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            InternalServerError:
              summary: An internal server error response object
              dataValue:
                type: https://example.com/errors/internal-server-error
                title: Internal Server Error
                status: 500
                detail: An unexpected error occurred.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred.
    NotFound:
      description: Not Found
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            NotFound:
              summary: A not found response object
              dataValue:
                type: https://example.com/errors/not-found
                title: Not Found
                status: 404
                detail: The requested resource was not found.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found.
    TooManyRequests:
      description: Too Many Requests
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
        Retry-After:
          $ref: '#/components/headers/Retry-After'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            TooManyRequests:
              summary: A too many requests response object
              dataValue:
                type: https://example.com/errors/too-many-requests
                title: Too Many Requests
                status: 429
                detail: You have exceeded the rate limit.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/too-many-requests
            title: Too Many Requests
            status: 429
            detail: You have exceeded the rate limit.
    Unauthorized:
      description: Unauthorized
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            Unauthorized:
              summary: An unauthorized response object
              dataValue:
                type: https://example.com/errors/unauthorized
                title: Unauthorized
                status: 401
                detail: You do not have the necessary permissions.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/unauthorized
            title: Unauthorized
            status: 401
            detail: You do not have the necessary permissions.
  examples:
...
    GetBookingsResponse200ApplicationJson:
      description: A list of bookings
      dataValue:
        data:
          - id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
            trip_id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
            passenger_name: John Doe
            has_bicycle: true
            has_dog: true
            status: confirmed
            price: 50
            currency: eur
            expires_at: '2026-10-02T11:04:31Z'
          - id: b2e783e1-c824-4d63-b37a-d8d698862f1d
            trip_id: b2e783e1-c824-4d63-b37a-d8d698862f1d
            passenger_name: Jane Smith
            has_bicycle: false
            has_dog: false
            status: confirmed
            price: 50
            currency: eur
            expires_at: '2026-10-02T11:04:31Z'
        links:
          self: https://api.example.com/bookings
          next: https://api.example.com/bookings?page=2
    CreateBookingResponse201ApplicationJson:
      description: A booking created
      dataValue:
        id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
        trip_id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
        passenger_name: John Doe
        has_bicycle: true
        has_dog: true
        status: pending_payment
        price: 50
        currency: eur
        expires_at: '2026-10-02T11:04:31Z'
        links:
          self: https://api.example.com/bookings/efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
    GetBookingResponse200ApplicationJson:
      description: A booking object
      dataValue:
        id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
        trip_id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
        passenger_name: John Doe
        has_bicycle: true
        has_dog: true
        status: confirmed
        price: 50
        currency: eur
        expires_at: '2026-10-02T11:04:31Z'
        links:
          self: https://api.example.com/bookings/1725ff48-ab45-4bb5-9d02-88745177dedb
    RequestBodyBookingApplicationJson:
      description: A booking object
      dataValue:
        trip_id: ea399ba1-6d95-433f-92d1-83f67b775594
        passenger_name: John Doe
        has_bicycle: true
        has_dog: true